Fix two warnings from the previous patch.

+++ src/resmom/linux/mom_mach.c
@@ -253,7 +253,7 @@ void proc_get_btime(void)
 
   fclose(fp);
 
-  sprintf(log_buffer, "DRIFT debug: getting btime, setting linux_time to %ld", linux_time);
+  sprintf(log_buffer, "DRIFT debug: getting btime, setting linux_time to %ld", (long)linux_time);
   log_event(PBSEVENT_DEBUG, PBS_EVENTCLASS_MOM, "linux_time", log_buffer);
 
   return;
@@ -4764,7 +4764,7 @@ void scan_non_child_tasks(void)
       }
       if(log_drift_event || LOGLEVEL >= 7)
       {
-          sprintf(log_buffer, "DRIFT debug: comparing linux_time %ld; job_start_time %ld and session_start_time[%ld] %ld: difference %d",
+          sprintf(log_buffer, "DRIFT debug: comparing linux_time %u; job_start_time %ld and session_start_time[%ld] %ld: difference %d",
           linux_time,
           job_start_time,
           job_session_id,
